KeywordNotifier/requirements.txt
fuzzywuzzy==0.18.0
Levenshtein==0.25.1
pyaes==1.6.1
pyasn1==0.6.0
python-decouple==3.8
python-Levenshtein==0.25.1
rapidfuzz==3.9.6
rsa==4.9
Telethon==1.36.0
